Introduce feature directories in tests
This sets up a directory structure where an entire directory gets a
feature mark. This makes it easy to skip entire feature tests.
