+++ tests/feature/iop/test_ingress.py
@@ -1,8 +1,3 @@
-import pytest
-
-pytestmark = pytest.mark.feature("iop")
-
-
 def test_ingress_service(server):
     service = server.service("iop-core-ingress")
     assert service.is_running
