+++ tests/feature/iop/test_puptoo.py
@@ -1,8 +1,3 @@
-import pytest
-
-pytestmark = pytest.mark.feature("iop")
-
-
 def test_puptoo_service(server):
     service = server.service("iop-core-puptoo")
     assert service.is_running
