+++ tests/feature/iop/test_vulnerability.py
@@ -1,8 +1,3 @@
-import pytest
-
-pytestmark = pytest.mark.feature("iop")
-
-
 def test_vulnerability_manager_service(server):
     service = server.service("iop-service-vuln-manager")
     assert service.is_running
