+++ tests/feature/iop/test_yuptoo.py
@@ -1,8 +1,3 @@
-import pytest
-
-pytestmark = pytest.mark.feature("iop")
-
-
 def test_yuptoo_service(server):
     service = server.service("iop-core-yuptoo")
     assert service.is_running
